Separate public APIs and implementation files in different packages
